Hooked up inventory system to Inventory Menu
Inventory System is hooked up to the inventory menu, now. Currently, it will just display the name of the item in the corresponding box along with the count of the item. In the end, I would like it to instead display an icon for the object but at the very least the logic needed is already there. I still need to write an onInteract Script for items that you can pickup and then make sure the functions they call operate properly.