Tested Inventory System and Added itemPickup
Tested the variations of adding items to inventory to make sure they all work. I also finalized the item pickup function that Logan started. I added a part of it so that you can specify if the pickup count of an item is affected by the current player score. The onInteract function wouldn't work on the handgun that Logan had in there but I duplicated the note books and replaced the script with this new one and it worked fine so, whatever is causing it, isn't on my end.