Added Objective to Inventory. Started HUD text.
Added an objective part to the inventory screen. Started working on the HUD section to display info to the player. It isn't set up yet but whenever a message is sent to it, the plan is that it'll fade in, stay on screen for a few seconds, and then fade out.